# server/documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Unicamp Tickets
  description: Tickets Tool for Unicamp
  contact:
    name: 'Tales Abdo dos Santos'
    email: talesabdo@outlook.com
  version: 1.0.0
paths:
  /user/new:
    post:
      operationId: insertUser
      summary: Create a new user
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/userNameParam'
        - $ref: '#/components/parameters/emailParam'
        - $ref: '#/components/parameters/passwordParam'
        - $ref: '#/components/parameters/imageParam'
      responses:
        200:
          description: Return OK message
  /user/update:
    put:
      operationId: updateUser
      summary: Update an especific user
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/userIdParam'
        - $ref: '#/components/parameters/userNameParam'
        - $ref: '#/components/parameters/passwordParam'
        - $ref: '#/components/parameters/imageParam'
      responses:
        200:
          description: Return OK message
  /user/delete/{id}:
    delete:
      operationId: deleteUser
      summary: Delete an especific user
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/userIdParam'
      responses:
        200:
          description: Return OK message
  /user/search:
    post:
      operationId: searchUsers
      summary: Search for users with similar name or email
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/typedTextParam'
      responses:
        200:
          description: Return an array with all matching users
  /user/authenticate:
    post:
      operationId: authenticateUser
      summary: Authenticate the inserted credentials
      tags:
        - User
      parameters:
        - $ref: '#/components/parameters/emailParam'
        - $ref: '#/components/parameters/passwordParam'
      responses:
        200:
          description: Return credential's existence
  /user/support/new/{id}:
    put:
      operationId: insertSupportUser
      summary: Give support user access to an existing common user
      tags:
        - Support User
      parameters:
        - $ref: '#/components/parameters/userIdParam'
      responses:
        200:
          description: Return OK message
  /user/support/delete/{id}:
    delete:
      operationId: deleteSupportUser
      summary: Remove support user access from an existing common user
      tags:
        - Support User
      parameters:
        - $ref: '#/components/parameters/userIdParam'
      responses:
        200:
          description: Return OK message
...
components:
  parameters:
    userIdParam:
      name: userId
      in: path
      required: true
      description: ID from an user
      schema:
        type: integer
    emailParam:
      name: email
      in: query
      required: true
      description: Email from an user
      schema:
        type: string
    userNameParam:
      name: name
      in: query
      required: true
      description: Name from an user
      schema:
        type: string
    passwordParam:
      name: password
      in: query
      required: true
      description: Password from an user
      schema:
        type: string
    imageParam:
      name: image
      in: query
      required: true
      description: Path from user image
      schema:
        type: string
    typedTextParam:
      name: typedText
      in: query
      required: true
      description: String used for search by name and email
      schema:
        type: string
...